Reviewed and made some very minor changes
Reviewed and tested the lab instructions on the icp2103 cluster in Raleigh. 
Made some very minor updates to punctuation and grammar for readability, and to conform to IBM style. 
Changed the name of chart ibm-mq-advanced-dev to ibm-mqadvanced-server-dev (to match what I saw listed in Catalog). 
At ln 241, added more detail in the note about Services, because I thought it needed some clarification.